Move plugins into own directory
Fix lint
