+++ modules/plugins/check-props-plugin.js
@@ -1,3 +1,5 @@
+/* @flow */
+
 var checkProps = function () {};
 
 if (process.env.NODE_ENV !== 'production') {
