+++ modules/plugins/merge-style-array-plugin.js
@@ -1,0 +1,10 @@
+/* @flow */
+
+// Convenient syntax for multiple styles: `style={[style1, style2, etc]}`
+// Ignores non-objects, so you can do `this.state.isCool && styles.cool`.
+var mergeStyleArrayPlugin = function ({style, mergeStyles}) {
+  var newStyle = Array.isArray(style) ? mergeStyles(style) : style;
+  return {style: newStyle};
+};
+
+module.exports = mergeStyleArrayPlugin;
