+++ modules/plugins/prefix-plugin.js
@@ -1,0 +1,10 @@
+/* @flow */
+
+var Prefixer = require('../prefixer');
+
+var prefixPlugin = function ({component, style}) {
+  var newStyle = Prefixer.getPrefixedStyle(component, style);
+  return {style: newStyle};
+};
+
+module.exports = prefixPlugin;
