+++ modules/plugins/resolve-interaction-styles-plugin.js
