+++ modules/plugins/resolve-media-queries-plugin.js
